// src/hotels/schemas/location.schema.ts
import { Prop, Schema } from '@nestjs/mongoose';

@Schema()
export class Location {
  @Prop({ type: String, enum: ['Point'] })
  type: string;

  @Prop({ index: '2dsphere' })
  coordinates: number[];

  formattedAddress: string;

  city: string;

  state: string;

  zipcode: string;

  country: string;
}